feat(privacy): differential privacy with budgets, ledger and receipts
Every privacy-relevant release now has a measured cost, and Encompute
refuses releases that exceed the approved budget.

- Budgets per asset: `privacy unit "patient" epsilon 3.0 delta 1e-6`.
  Python: `asset(..., privacy="strong", unit="patient")` or DP(e, d),
  with named levels standard, strong and maximum.
- The compiler detects release boundaries: revealing a budgeted asset
  needs a DP mechanism (ENC2203). Sealed values cost nothing.
- Discrete Gaussian noise on the secure-aggregation sum
  (`dp discrete_gaussian clip_norm C noise_multiplier z`; Python
  secure_aggregate(..., privacy=...)). Parties L2-clip their
  contributions; the noise is exact integer noise (a port of the CKS 2020
  sampler) from a CSPRNG.
- zCDP accounting with the CKS conversion, deterministic through libm
  and rounded up.
- encompute-privacy: a hash-chained, file-locked, persistent ledger per
  asset with reserve-then-commit releases (no unaccounted release, no
  concurrent double spend) and signed PrivacyReceipts. Owners verify
  their ledger and last checkpoint before contributing, and refuse
  over-budget rounds (ENC2201) and rolled-back or reset ledgers
  (ENC2202).
- PrivacyPolicyId (encprivacy1:) is bound into the ExecutionSpec,
  aggregation plan, ledgers, receipts and coordinator attestation.
  Parties can require an attested coordinator bound to the privacy
  configuration.
- CLI: `privacy budget`, `explain --ledger`, `aggregate serve --ledger`;
  --state keeps ledger checkpoints. Errors ENC2201-ENC2204.
- examples/private_federated_training: 12 rounds permitted, the 13th
  denied. ADR-013.

+++ crates/encompute-runtime/tests/aggregation.rs
@@ -132,7 +132,9 @@ fn three_hospitals_over_http() {
     let codec = spec.plan.codec;
     for j in 0..len {
         let clear: f64 = (0..3).map(|i| gradient(i, len)[j]).sum();
-        let encoded: u64 = (0..3).map(|i| codec.encode(gradient(i, len)[j])).sum();
+        let encoded: i64 = (0..3)
+            .map(|i| codec.encode(gradient(i, len)[j]) as i64)
+            .sum();
         assert_eq!(aggregate.encoded_sum[j], encoded);
         assert_eq!(aggregate.values[j], codec.decode_sum(encoded, 3));
         assert!((aggregate.values[j] - clear).abs() <= 3.0 * codec.resolution() + 1e-12);
@@ -270,9 +272,9 @@ fn dropouts_within_the_threshold() {
     );
     let codec = spec.plan.codec;
     for j in 0..64 {
-        let enc: u64 = [0, 2, 4]
+        let enc: i64 = [0, 2, 4]
             .iter()
-            .map(|&i| codec.encode(gradient(i, 64)[j]))
+            .map(|&i| codec.encode(gradient(i, 64)[j]) as i64)
             .sum();
         assert_eq!(agg.encoded_sum[j], enc);
     }
@@ -298,9 +300,9 @@ fn mean_divides_by_the_contributors() {
     let (agg, _) = run(&mut c, &mut parts, &[None, None, Some(2), None]).unwrap();
     let codec = spec.plan.codec;
     for j in 0..8 {
-        let enc: u64 = [0, 1, 3]
+        let enc: i64 = [0, 1, 3]
             .iter()
-            .map(|&i| codec.encode(gradient(i, 8)[j]))
+            .map(|&i| codec.encode(gradient(i, 8)[j]) as i64)
             .sum();
         assert_eq!(agg.values[j], codec.decode_sum(enc, 3) / 3.0);
     }
